+++ src/gravity-client/AbstractGravityClient.ts
@@ -32,9 +32,10 @@ export interface GravityClientOptions {
   onPublish?: (userActions: ReadonlyArray<SessionUserAction>) => void
 }
 
-export type RecordingSettings = Pick<CollectorOptions, 'enableEventRecording' | 'enableVideoRecording'> & {
-  enableVideoAnonymization: boolean
-}
+export type RecordingSettings = Pick<
+  CollectorOptions,
+  'enableEventRecording' | 'enableVideoRecording' | 'enableVideoAnonymization'
+>
 
 export default abstract class AbstractGravityClient implements IGravityClient {
   private readonly sessionUserActionBuffer: DataBuffering<SessionUserAction, AddSessionUserActionsResponse>