+++ src/utils/completeOptions.ts
@@ -22,6 +22,7 @@ export default function completeOptions(options?: Partial<CollectorOptions>): Co
     rejectSession: DEFAULT_SESSION_REJECTION,
     enableEventRecording: true,
     enableVideoRecording: options.disableVideoRecording === undefined ? true : !options.disableVideoRecording,
+    enableVideoAnonymization: true,
   }
 
   const debugDefaultOptions = {